django-notifier/requirements.txt
# REST Api
Django==2.1.7
djangorestframework==3.8.2
djangorestframework_jwt==1.11.0
coreapi==2.3.0
openapi-codec==1.3.1
simplejson==3.9.0
django-filter==2.1.0
django-cors-headers==2.4.1
django-extensions==2.1.3

# PostGreSQL
psycopg2==2.7.3.2
# Multiple instances
gunicorn==19.7.0
# Static files
whitenoise==3.3.0


# Notification
channels==2.1.7
channels-redis==2.3.3